tokenizer: Disable language specific period handling.
For some reason, I am getting lots of periods not separated from the
preceding words at the end of sentence. It seems this is caused
somehow by the language-specific abbreviation handling mechanism.

Since the language-specific stuff is useless for our purposes anyway,
unconditionally split the period from the preceding word.